Add plant.rb class definition
This class represents a plant and encapsulates attributes such as bloom time/color, height, moisture and sun needed, and both the common and scientific names--as well as an array of communities in which the plant can be found.

Behaviors defined include returning arrays of related attributes (bloom info, moisture and sun requirements, and both names) as well as whether or not the plant is found/part of a given community.